+++ source/benoit/benoit/application/render_row_julia.rs
@@ -29,7 +29,7 @@ extern crate rug;
 use rug::Float;
 
 impl Application {
-	pub fn render_row(data: &mut [u32], y: u32, canvas_width: u32, canvas_height: u32, center_real: Float, center_imaginary: Float, zoom: Float, maximum_iteration_count: u32) {
+	pub fn render_row_julia(data: &mut [u32], y: u32, canvas_width: u32, canvas_height: u32, center_real: Float, center_imaginary: Float, zoom: Float, maximum_iteration_count: u32, julia_real: Float, julia_imaginary: Float) {
 		for x in 0x0..canvas_width {
 			let canvas_width  = Float::with_val(PRECISION, canvas_width);
 			let canvas_height = Float::with_val(PRECISION, canvas_height);
@@ -37,46 +37,40 @@ impl Application {
 			let x_float = Float::with_val(PRECISION, x);
 			let y_float = Float::with_val(PRECISION, y);
 
-			// Re(c) = (x-canvas_width/2)/canvas_width*4/zoom+Re(z)
-			let ca = {
+			// The Julia set of the Mandelbrot set is similar
+			// but not quite identical: The start value of (z)
+			// is now relative to the canvas and (c) is
+			// constant corresponds to a point in the
+			// Mandelbrot Set.
+
+			let ca = &julia_real;
+			let cb = &julia_imaginary;
+
+			// Re(z) = (x-canvas_width/2)*4/canvas_width/zoom+Re(z)
+			let mut za = {
 				let tmp0 = Float::with_val(PRECISION, &canvas_width / 2.0);
 
-				let mut ca = Float::with_val(PRECISION, &x_float - &tmp0);
-				ca /= &canvas_width;
-				ca *= 4.0;
-				ca /= &zoom;
-				ca += &center_real;
+				let mut za = Float::with_val(PRECISION, &x_float - &tmp0);
+				za *= 4.0;
+				za /= &canvas_width;
+				za /= &zoom;
+				za += &center_real;
 
-				ca
+				za
 			};
 
-			// Im(c) = (x-canvas_height/2)/canvas_height*4/zoom+Im(z)
-			let cb = {
+			// Im(Z) = (x-canvas_height/2)*4/canvas_height/zoom+Im(z)
+			let mut zb = {
 				let tmp0 = Float::with_val(PRECISION, &canvas_height / 2.0);
 
-				let mut cb = Float::with_val(PRECISION, &y_float - &tmp0);
-				cb /= &canvas_height;
-				cb *= 4.0;
-				cb /= &zoom;
-				cb += &center_imaginary;
+				let mut zb = Float::with_val(PRECISION, &y_float - &tmp0);
+				zb *= 4.0;
+				zb /= &canvas_height;
+				zb /= &zoom;
+				zb += &center_imaginary;
 
-				cb
+				zb
 			};
-
-			// Formaly, the initial condition that
-			//
-			// z = 0
-			//
-			// may be skipped as the first two iterations will
-			// always be
-			//
-			// z0 = 0
-			// z1 = z0^2+c = 0^2+c = 0+c = c
-			//
-			// That is, the result of the second iteration will
-			// always be the value of (c).
-			let mut za = Float::with_val(PRECISION, &ca);
-			let mut zb = Float::with_val(PRECISION, &cb);
 
 			let mut iteration_count: u32 = 0x0;
 			while {
@@ -84,27 +78,23 @@ impl Application {
 				square_distance <= 4.0 && iteration_count < maximum_iteration_count
 			} {
 				{
-					// The Mandelbrot Set (M) is defined as the set of values in the complex plane where the iterating function
+					// The overall iterations of the Julia of M are
+					// identical to those of M:
 					//
 					// z = z^2+c
 					//
-					// stays bounded: I.e. the absolute value of (z) stays bounded:
-					//
-					// abs(z) = sqrt(Re(z)^2+Im(z)^2) <= 2^2 = 4
+					// with only the initial value of (z) and (c)
+					// differing.
 
 					let za_temporary = Float::with_val(PRECISION, &za);
 
-					// We can calculate the square of a complex number (z) as:
-					//
-					// z^2 = (a+ib)^2 = (a+ib)(a+ib) = a^2+iab+iab-b^2 = a^2-b^2+2iab
-
 					za = za.square();
 					za -= &zb * &zb;
-					za += &ca;
+					za += ca;
 
 					zb *= &za_temporary;
 					zb *= 2.0;
-					zb += &cb;
+					zb += cb;
 				}
 
 				iteration_count += 0x1;